Update the tests
