+++ core/src/main/scala/munit/CatsEffectFunFixtures.scala
