+++ core/src/test/scala/munit/CatsEffectAssertionsSyntaxSpec.scala
