+++ core/src/test/scala/munit/CatsEffectFixturesSpec.scala
@@ -17,6 +17,8 @@
 package munit
 
 import cats.effect.{IO, Resource}
+import munit.catseffect.ResourceFixture.FixtureNotInstantiatedException
+
 import scala.concurrent.duration._
 
 class CatsEffectFixturesSpec extends CatsEffectSuite with CatsEffectAssertions {
@@ -24,7 +26,7 @@ class CatsEffectFixturesSpec extends CatsEffectSuite with CatsEffectAssertions {
   @volatile var acquired: Int = 0
   @volatile var released: Int = 0
 
-  val fixture = UnsafeResourceSuiteLocalDeferredFixture(
+  val fixture = ResourceSuiteLocalFixture(
     "fixture",
     Resource.make(
       IO.sleep(1.millis) *> IO {
@@ -39,6 +41,11 @@ class CatsEffectFixturesSpec extends CatsEffectSuite with CatsEffectAssertions {
     )
   )
 
+  val uninitializedFixture = ResourceSuiteLocalFixture(
+    "uninitialized-fixture",
+    Resource.make(IO.unit)(_ => IO.unit)
+  )
+
   override def munitFixtures = List(fixture)
 
   override def beforeAll(): Unit = {
@@ -48,15 +55,22 @@ class CatsEffectFixturesSpec extends CatsEffectSuite with CatsEffectAssertions {
 
   override def afterAll(): Unit = {
     assertEquals(acquired, 1)
-    // assertEquals(released, 1) // Release is async, no way to check
+    assertEquals(released, 1)
   }
 
   test("first test") {
-    fixture().assertEquals(())
+    IO(fixture()).assertEquals(())
   }
 
   test("second test") {
-    fixture().assertEquals(())
+    IO(fixture()).assertEquals(())
+  }
+
+  test("throws exception") {
+    IO(uninitializedFixture())
+      .interceptMessage[FixtureNotInstantiatedException](
+        "The fixture `uninitialized-fixture` was not instantiated. Override `munitFixtures` and include a reference to this fixture."
+      )
   }
 
 }
