+++ core/src/test/scala/munit/ExampleSuite.scala
@@ -41,7 +41,7 @@ class ExampleSuite extends CatsEffectSuite {
 
   import cats.effect.std.Dispatcher
 
-  val dispatcher = ResourceFixture(Dispatcher[IO])
+  val dispatcher = ResourceFunFixture(Dispatcher[IO])
 
   dispatcher.test("resources can be lifted to munit fixtures") { dsp =>
     dsp.unsafeRunAndForget(IO(42))
